+++ history.rst
@@ -6,6 +6,17 @@ by Gerold Penz 2013-2014
 
 
 =============
+Version 0.4.2
+=============
+
+2014-09-12
+
+- HttpClient: New parameters added:
+  - additional_headers: Posibility to add additional header items.
+  - content_type: Posibility to change the content-type header.
+
+
+=============
 Version 0.4.1
 =============